refactor(storefront): extract CatalogTemplate component to eliminate duplicate code and integrate ActiveFilterChips
